Revert "Added zend-hydrator v3 support [No BC Break]"
