+++ composer.json
@@ -17,7 +17,7 @@
         "php": "^7.1",
         "ext-ctype": "*",
         "doctrine/common": "^2.8",
-        "zendframework/zend-hydrator": "^2.3 || ^3.0",
+        "zendframework/zend-hydrator": "^2.3",
         "zendframework/zend-stdlib": "^3.1"
     },
     "require-dev": {
@@ -27,17 +27,11 @@
         "zendframework/zend-servicemanager": "^3.3"
     },
     "autoload": {
-        "files": [
-            "src/_autoload.php"
-        ],
         "psr-4": {
             "Doctrine\\Zend\\Hydrator\\": "src/"
         }
     },
     "autoload-dev": {
-        "files": [
-            "test/_autoload.php"
-        ],
         "psr-4": {
             "DoctrineTest\\Zend\\Hydrator\\": "test/"
         }
