+++ src/DoctrineObject.php
@@ -17,10 +17,7 @@ use Zend\Hydrator\AbstractHydrator;
 use Zend\Hydrator\Filter\FilterProviderInterface;
 use Zend\Stdlib\ArrayUtils;
 
-/**
- * @internal Only for internal use to deliver support for zend-hydrator v2 and v3.
- */
-abstract class DoctrineObjectInternal extends AbstractHydrator
+class DoctrineObject extends AbstractHydrator
 {
     /**
      * @var ObjectManager
@@ -48,6 +45,18 @@ abstract class DoctrineObjectInternal extends AbstractHydrator
     protected $defaultByReferenceStrategy = AllowRemoveByReference::class;
 
     /**
+     * @param ObjectManager $objectManager The ObjectManager to use
+     * @param bool $byValue If set to true, hydrator will always use entity's public API
+     */
+    public function __construct(ObjectManager $objectManager, $byValue = true)
+    {
+        parent::__construct();
+
+        $this->objectManager = $objectManager;
+        $this->byValue = (bool) $byValue;
+    }
+
+    /**
      * @return string
      */
     public function getDefaultByValueStrategy()
@@ -57,7 +66,7 @@ abstract class DoctrineObjectInternal extends AbstractHydrator
 
     /**
      * @param string $defaultByValueStrategy
-     * @return $this
+     * @return DoctrineObject
      */
     public function setDefaultByValueStrategy($defaultByValueStrategy)
     {
@@ -75,7 +84,7 @@ abstract class DoctrineObjectInternal extends AbstractHydrator
 
     /**
      * @param string $defaultByReferenceStrategy
-     * @return $this
+     * @return DoctrineObject
      */
     public function setDefaultByReferenceStrategy($defaultByReferenceStrategy)
     {
@@ -86,12 +95,10 @@ abstract class DoctrineObjectInternal extends AbstractHydrator
     /**
      * Extract values from an object
      *
-     * @internal Used in v2 and v3 proxies.
-     *
      * @param  object $object
      * @return array
      */
-    protected function extractInternal($object)
+    public function extract($object)
     {
         $this->prepare($object);
 
@@ -105,13 +112,11 @@ abstract class DoctrineObjectInternal extends AbstractHydrator
     /**
      * Hydrate $object with the provided $data.
      *
-     * @internal Used in v2 and v3 proxies.
-     *
      * @param  array $data
      * @param  object $object
      * @return object
      */
-    protected function hydrateInternal(array $data, $object)
+    public function hydrate(array $data, $object)
     {
         $this->prepare($object);
 
@@ -159,7 +164,7 @@ abstract class DoctrineObjectInternal extends AbstractHydrator
 
                 $strategy = $this->getStrategy($association);
 
-                if (! $strategy instanceof Strategy\AbstractCollectionStrategyInternal) {
+                if (! $strategy instanceof Strategy\AbstractCollectionStrategy) {
                     throw new InvalidArgumentException(
                         sprintf(
                             'Strategies used for collections valued associations must inherit from '
@@ -253,7 +258,7 @@ abstract class DoctrineObjectInternal extends AbstractHydrator
      *
      * @inheritdoc
      */
-    protected function hydrateValueInternal($name, $value, $data = null)
+    public function hydrateValue($name, $value, $data = null)
     {
         $value = parent::hydrateValue($name, $value, $data);
