+++ src/Filter/PropertyName.php
@@ -9,10 +9,8 @@ use Zend\Hydrator\Filter\FilterInterface;
 /**
  * Provides a filter to restrict returned fields by whitelisting or
  * blacklisting property names.
- *
- * @internal
  */
-abstract class PropertyNameInternal implements FilterInterface
+class PropertyName implements FilterInterface
 {
     /**
      * The propteries to exclude.
@@ -38,5 +36,12 @@ abstract class PropertyNameInternal implements FilterInterface
         $this->properties = is_array($properties)
             ? $properties
             : [$properties];
+    }
+
+    public function filter($property)
+    {
+        return in_array($property, $this->properties)
+            ? ! $this->exclude
+            : $this->exclude;
     }
 }
