+++ src/Strategy/AllowRemoveByReference.php
@@ -10,16 +10,13 @@ namespace Doctrine\Zend\Hydrator\Strategy;
  * collection initially contains elements A and B, and that the new collection contains elements B and C, then the
  * final collection will contain elements B and C (while element A will be asked to be removed).
  * This strategy is by reference, this means it won't use public API to add/remove elements to the collection
- *
- * @internal
  */
-trait AllowRemoveByReferenceTrait
+class AllowRemoveByReference extends AbstractCollectionStrategy
 {
     /**
-     * @param mixed $value
-     * @return mixed
+     * {@inheritDoc}
      */
-    private function hydrateInternal($value)
+    public function hydrate($value)
     {
         $collection = $this->getCollectionFromObjectByReference();
         $collectionArray = $collection->toArray();
