+++ src/Strategy/DisallowRemoveByReference.php
@@ -10,16 +10,13 @@ namespace Doctrine\Zend\Hydrator\Strategy;
  * instance, if the collection initially contains elements A and B, and that the new collection contains elements B
  * and C, then the final collection will contain elements A, B and C.
  * This strategy is by reference, this means it won't use the public API to remove elements
- *
- * @internal
  */
-trait DisallowRemoveByReferenceTrait
+class DisallowRemoveByReference extends AbstractCollectionStrategy
 {
     /**
-     * @param mixed $value
-     * @return mixed
+     * {@inheritDoc}
      */
-    private function hydrateInternal($value)
+    public function hydrate($value)
     {
         $collection = $this->getCollectionFromObjectByReference();
         $collectionArray = $collection->toArray();
