--- src/Strategy/DisallowRemoveByReferenceV2.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Doctrine\Zend\Hydrator\Strategy;
-
-/**
- * @internal
- */
-class DisallowRemoveByReferenceV2 extends AbstractCollectionStrategy
-{
-    use DisallowRemoveByReferenceTrait;
-
-    /**
-     * @param mixed $value
-     * @return mixed
-     */
-    public function hydrate($value)
-    {
-        return $this->hydrateInternal($value);
-    }
-}
