--- src/Strategy/DisallowRemoveByValueV2.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Doctrine\Zend\Hydrator\Strategy;
-
-/**
- * @internal
- */
-class DisallowRemoveByValueV2 extends AbstractCollectionStrategy
-{
-    use DisallowRemoveByValueTrait;
-
-    /**
-     * @param mixed $value
-     * @return mixed
-     */
-    public function hydrate($value)
-    {
-        return $this->hydrateInternal($value);
-    }
-}
