+++ test/Assets/SimpleStrategy.php
@@ -6,14 +6,14 @@ namespace DoctrineTest\Zend\Hydrator\Assets;
 
 use Zend\Hydrator\Strategy\StrategyInterface;
 
-class SimpleStrategyV3 implements StrategyInterface
+class SimpleStrategy implements StrategyInterface
 {
-    public function extract($value, ?object $object = null)
+    public function extract($value)
     {
         return 'modified while extracting';
     }
 
-    public function hydrate($value, ?array $data)
+    public function hydrate($value)
     {
         return 'modified while hydrating';
     }
